tests/wpt/web-platform-tests/IndexedDB/idbdatabase_deleteObjectStore2.htm

<!DOCTYPE html>
<title>IDBDatabase.deleteObjectStore() - attempt to remove an object store outside of a version change transaction </title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support.js"></script>

<script>

var t = async_test(),
    open_rq = createdb(t);

open_rq.onupgradeneeded = function(e)
{
    var db = e.target.result,
        objStore = db.createObjectStore("delete_outside");

    e.target.transaction.oncomplete = t.step_func(function (e)
    {
        assert_throws('InvalidStateError',
            function() { db.deleteObjectStore("delete_outside"); });
        t.done();
    });
}

</script>


<div id="log"></div>
